// TOORLOP/src/types.ts
export interface User {
  id: string;
  name: string;
  avatar: string;
  coverPhoto?: string;
  online?: boolean;
  /**
   * Indicates if the user has locked their profile.
   * If true, strictly restricts content visibility (posts, photos, videos) to friends only.
   */
  isLocked?: boolean;
}
// ...
export interface ProductComment {
  id: string;
  user: User;
  text: string;
  timestamp: string;
}

export interface Product {
  id: string;
  title: string;
  price: number;
  currency: string;
  category: string;
  image: string;
  images?: string[]; // Added support for multiple images
  location: string;
  seller: User;
  description: string;
  condition: 'new' | 'used_good' | 'used_fair';
  date: string;
  isSaved?: boolean;
  timestamp: number; // For sorting
  comments?: ProductComment[];
}